Adding 'variables view' to data grid
Salstat now has a ‘variables view’, similar to SPSS. This is a tab
whereby users can view the metadata of each variable. It differs from
SPSS in that variables are shown in columns rather than rows. This is
to preserve the visual ’language / grammar’ that a column represents a
variable.